Сделаны правки по конвертации значений, убраны ненужные преобразования

+++ data/src/main/java/com/golandcoinc/data/db/dao/AppDao.kt
@@ -42,7 +42,7 @@ interface AppDao {
     fun setRecommendedSpeed(speed: SpeedDBModel)
 
     @Query("SELECT speed FROM SpeedValue WHERE id = 0")
-    fun getRecommendedSpeed(): Int?
+    fun getRecommendedSpeed(): Float?
 
     @Query("DELETE FROM SpeedValue")
     fun clearRecommendedSpeed()
